+++ core/image.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2008-2019 the MRtrix3 contributors.
+/* Copyright (c) 2008-2021 the MRtrix3 contributors.
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -473,7 +473,7 @@ namespace MR
       int64_t offset = 0;
       out << "file: ";
       if (single_file) {
-        offset = out.tellp() + int64_t(18);
+        offset = int64_t(out.tellp()) + int64_t(18);
         offset += ((4 - (offset % 4)) % 4);
         out << ". " << offset << "\nEND\n";
       }
